// src/app/core/models/school.profile.model.ts
export class SchoolProfileModel {
    school_id: number;
    school_name: string;
    school_logo_image: string;
    school_address: string;
    school_service_desc: string;
    school_link_youtube: string;
    school_website_url: string;   
    school_cover_image: string;
    //add phone number
    school_phone_number: string;
}